Improved time zone check.
Previously, the zone check compared time.Local and time.UTC. However,
when this was tested on a machine that uses UTC locally, it became
apparent that the two pointers can differ while having the same time
zone. A more useful check is to look up the offsets for the time zones
and compare those --- they'll return the same times.